refactor(view): Move menu model in constants to new package
